+++ package.json
@@ -6,17 +6,18 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "antd": "^5.1.6",
     "axios": "^1.1.3",
+    "firebase": "^9.6.7",
+    "firebase-tools": "^10.2.2",
     "lint-staged": "^11.2.6",
+    "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-cookie": "^4.1.1",
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.6.2",
     "react-scripts": "5.0.1",
-    "web-vitals": "^1.0.1",
-    "firebase": "^9.6.7",
-    "firebase-tools": "^10.2.2",
-    "prop-types": "^15.8.1"
+    "web-vitals": "^1.0.1"
   },
   "optionalDependencies": {
     "babel-loader": "8.1.0"
