+++ src/components/Login/Login.module.css
@@ -1,0 +1,43 @@
+.logo-container {
+  position: absolute;
+  display: flex;
+  flex-direction: row;
+  column-gap: 15px;
+  font-size: 20px;
+  padding: 0.5% 1.5%;
+}
+
+.logo-container h1 {
+  font-weight: 400;
+  font-size: 18px;
+}
+
+.container {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100vw;
+  height: 100vh;
+}
+
+.container h1 {
+  font-weight: bold;
+}
+
+.login-form {
+  width: 280px;
+}
+
+.login-form label {
+  font-weight: 500;
+  font-size: 20px !important;
+  line-height: 22px;
+}
+
+.login-form button {
+  border-radius: 3px;
+  height: 40px;
+  font-weight: 500;
+  left: calc(50% - 35.875px);
+}
